+++ addon/plugins/decision-plugin/commands/index.ts
@@ -1,0 +1,4 @@
+export { default as insertTitle } from './insert-title';
+export { default as insertArticleContainer } from './insert-article-container';
+export { default as insertDescription } from './insert-description';
+export { default as insertMotivation } from './insert-motivation';