+++ addon/plugins/table-of-contents-plugin/nodes/table-of-contents.ts
@@ -26,6 +26,17 @@ export const emberNodeConfig: (
     toDOM(node) {
       const { entries } = node.attrs;
 
+      if (!entries) {
+        return [
+          'div',
+          {
+            'data-ember-node': 'table-of-contents',
+            class: 'table-of-contents',
+          },
+          ['h3', {}, 'Table Of Contents'],
+        ];
+      }
+
       return [
         'div',
         { 'data-ember-node': 'table-of-contents', class: 'table-of-contents' },