+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Use plugin configuration instead of ember environment in all the plugins
 
+### Added
+- validation plugin
+ 
+### Deprecated
+- the decision-plugin card component is now deprecated in favor of the host app's choice of insert button
+
 ## [5.0.1] - 2023-04-07
 ### Fixed
 - fix deprecated use of modifier without specifying eagerness